requirements.txt
Flask==3.1.3
python-dotenv==1.2.1
openai==2.32.0
gunicorn==21.2.0
